backend/utils/auth_utils.py
from fastapi import Header, HTTPException
import jwt, os
JWT_SECRET = os.getenv('JWT_SECRET', 'change_me')
JWT_ALG = 'HS256'
def get_current_user(authorization: str = Header(None)):
    if not authorization or not authorization.startswith("Bearer "):
        raise HTTPException(status_code=401, detail="Missing or invalid Authorization header")
    token = authorization.split()[1]
    try:
        payload = jwt.decode(token, JWT_SECRET, algorithms=[JWT_ALG])
        return payload.get("sub")
    except Exception:
        raise HTTPException(status_code=401, detail="Invalid token")
