backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
openai
bcrypt
PyJWT
python-multipart
